+++ src/mailman/app/tests/test_notifications.py
@@ -32,7 +32,8 @@ from mailman.config import config
 from mailman.interfaces.languages import ILanguageManager
 from mailman.interfaces.member import MemberRole
 from mailman.interfaces.usermanager import IUserManager
-from mailman.testing.helpers import get_queue_messages, subscribe, set_preferred
+from mailman.testing.helpers import (
+    get_queue_messages, set_preferred, subscribe)
 from mailman.testing.layers import ConfigLayer
 from mailman.utilities.datetime import now
 from zope.component import getUtility
@@ -145,86 +146,71 @@ Welcome to the Test List mailing list.
         self.assertEqual(len(messages), 0)
 
     def test_member_susbcribed_address_has_display_name(self):
-        address = getUtility(IUserManager).create_address('anne@example.com',
-            'Anne Person')
+        address = getUtility(IUserManager).create_address(
+            'anne@example.com', 'Anne Person')
         address.verified_on = now()
         self._mlist.subscribe(address)
-        messages = get_queue_messages('virgin')
+        messages = get_queue_messages('virgin', expected_count=1)
         message = messages[0].msg
-        self.assertEqual(message['to'],
-            'Anne Person <anne@example.com>')
+        self.assertEqual(message['to'], 'Anne Person <anne@example.com>')
 
     def test_member_subscribed_address_has_no_display_name(self):
-        address = getUtility(IUserManager).create_address('anne@example.com',
-            '')
+        address = getUtility(IUserManager).create_address('anne@example.com')
         address.verified_on = now()
         self._mlist.subscribe(address)
-        messages = get_queue_messages('virgin')
+        messages = get_queue_messages('virgin', expected_count=1)
         message = messages[0].msg
-        self.assertEqual(message['to'],
-            'anne@example.com')
+        self.assertEqual(message['to'], 'anne@example.com')
 
     def test_member_is_user_and_has_display_name(self):
-        user = getUtility(IUserManager).create_user('anne@example.com',
-            'Anne Person')
+        user = getUtility(IUserManager).create_user(
+            'anne@example.com', 'Anne Person')
         set_preferred(user)
         self._mlist.subscribe(user)
-        messages = get_queue_messages('virgin')
+        messages = get_queue_messages('virgin', expected_count=1)
         message = messages[0].msg
-        self.assertEqual(message['to'],
-            'Anne Person <anne@example.com>')
+        self.assertEqual(message['to'], 'Anne Person <anne@example.com>')
 
     def test_member_is_user_and_has_no_display_name(self):
-        user = getUtility(IUserManager).create_user('anne@example.com',
-            '')
+        user = getUtility(IUserManager).create_user('anne@example.com')
         set_preferred(user)
         self._mlist.subscribe(user)
-        messages = get_queue_messages('virgin')
+        messages = get_queue_messages('virgin', expected_count=1)
         message = messages[0].msg
-        self.assertEqual(message['to'],
-            'anne@example.com')
+        self.assertEqual(message['to'], 'anne@example.com')
 
     def test_member_has_linked_user_display_name(self):
-        user = getUtility(IUserManager).create_user('anne@example.com',
-            'Anne Person')
+        user = getUtility(IUserManager).create_user(
+            'anne@example.com', 'Anne Person')
         set_preferred(user)
-        self._mlist.subscribe(user)
-        address = getUtility(IUserManager).create_address('anne2@example.com',
-            '')
+        address = getUtility(IUserManager).create_address('anne2@example.com')
         address.verified_on = now()
         user.link(address)
         self._mlist.subscribe(address)
-        messages = get_queue_messages('virgin')
-        message = messages[1].msg
-        self.assertEqual(message['to'],
-            'Anne Person <anne2@example.com>')
+        messages = get_queue_messages('virgin', expected_count=1)
+        message = messages[0].msg
+        self.assertEqual(message['to'], 'Anne Person <anne2@example.com>')
 
     def test_member_has_no_linked_display_name(self):
-        user = getUtility(IUserManager).create_user('anne@example.com',
-            '')
+        user = getUtility(IUserManager).create_user('anne@example.com')
         set_preferred(user)
-        self._mlist.subscribe(user)
-        address = getUtility(IUserManager).create_address('anne2@example.com',
-            '')
+        address = getUtility(IUserManager).create_address('anne2@example.com')
         address.verified_on = now()
         user.link(address)
         self._mlist.subscribe(address)
-        messages = get_queue_messages('virgin')
-        message = messages[1].msg
-        self.assertEqual(message['to'],
-            'anne2@example.com')
+        messages = get_queue_messages('virgin', expected_count=1)
+        message = messages[0].msg
+        self.assertEqual(message['to'], 'anne2@example.com')
 
     def test_member_has_address_and_user_display_name(self):
-        user = getUtility(IUserManager).create_user('anne@example.com',
-            'Anne Person')
+        user = getUtility(IUserManager).create_user(
+            'anne@example.com', 'Anne Person')
         set_preferred(user)
-        self._mlist.subscribe(user)
-        address = getUtility(IUserManager).create_address('anne2@example.com',
-            'Anne Person 2')
+        address = getUtility(IUserManager).create_address(
+            'anne2@example.com', 'Anne X Person')
         address.verified_on = now()
         user.link(address)
         self._mlist.subscribe(address)
-        messages = get_queue_messages('virgin')
-        message = messages[1].msg
-        self.assertEqual(message['to'],
-            'Anne Person 2 <anne2@example.com>')
+        messages = get_queue_messages('virgin', expected_count=1)
+        message = messages[0].msg
+        self.assertEqual(message['to'], 'Anne X Person <anne2@example.com>')